requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
django-environ